+++ packages/terra-functional-testing/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Fixed
-  * Update specPath in BaseCompare to replace `node_modules` with `tests/wdio/__snapshot__`
+  * Update specPath in BaseCompare to replace `node_modules` with `tests/wdio`.
 
 ## 1.2.0 - (April 23, 2021)